+++ src/dune_rules/coq_lib.mli
@@ -43,5 +43,18 @@ module DB : sig
 
   val requires_for_user_written :
     t -> (Loc.t * Coq_lib_name.t) list -> lib list Or_exn.t
+
+  (** Reverse map from source names to a matching lib, plus the matching suffix;
+      this is only required because coqdep doesn't handle native dependencies,
+      so for a dependency [path/bar/foo.vo] we must correctly translate it to
+      its mangled Coq native form, which requires locating the library which a
+      source does belong to. The current matching function does return a triple
+      [lib,
+     prefix, name] for a given input path, or raises if not found.
+
+      Note that this may be useful in the future if we are able to implement a
+      [-modules] option in coqdep, so we could do the resolution
+      [file -> module] ourselves. *)
+  val module_of_source_file : t -> Path.Build.t -> lib * string list * string
 end
 with type lib := t