+++ src/dune_rules/coq_module.ml
@@ -37,38 +37,52 @@ let prefix x = x.prefix
 
 let name x = x.name
 
-let build_vo_dir ~obj_dir x =
-  List.fold_left x.prefix ~init:obj_dir ~f:Path.Build.relative
+let build_vo_dir ~obj_dir ~prefix =
+  List.fold_left prefix ~init:obj_dir ~f:Path.Build.relative
+
+let mangled_native_filename ~wrapper_name ~prefix ~name =
+  (* Split on dot is required! *)
+  let wrapper_split = String.split wrapper_name ~on:'.' in
+  "N" ^ String.concat ~sep:"_" (wrapper_split @ prefix @ [ name ])
 
 let cmxs_of_mod ~wrapper_name x =
-  let wrapper_split = String.split wrapper_name ~on:'.' in
   let native_base =
-    "N" ^ String.concat ~sep:"_" (wrapper_split @ x.prefix @ [ x.name ])
+    mangled_native_filename ~wrapper_name ~prefix:x.prefix ~name:x.name
   in
   [ native_base ^ ".cmi"; native_base ^ ".cmxs" ]
 
+let native_mangle_filename ~wrapper_name ~prefix ~name ~obj_dir ~ext =
+  let native_base = mangled_native_filename ~wrapper_name ~prefix ~name in
+  let vo_dir = build_vo_dir ~obj_dir ~prefix in
+  Path.Build.relative vo_dir (native_base ^ ext)
+
 let dep_file x ~obj_dir =
-  let vo_dir = build_vo_dir ~obj_dir x in
+  let vo_dir = build_vo_dir ~obj_dir ~prefix:x.prefix in
   Path.Build.relative vo_dir (x.name ^ ".v.d")
 
 type obj_files_mode =
   | Build
   | Install
 
+let native_obj_files_core x ~wrapper_name ~vo_dir ~install_vo_dir =
+  let cmxs_obj = cmxs_of_mod ~wrapper_name x in
+  List.map
+    ~f:(fun x ->
+      ( Path.Build.relative vo_dir x
+      , Filename.(concat (concat install_vo_dir ".coq-native") x) ))
+    cmxs_obj
+
 (* XXX: Remove the install .coq-native hack once rules can output targets in
    multiple subdirs *)
 let obj_files x ~wrapper_name ~mode ~obj_dir ~obj_files_mode =
-  let vo_dir = build_vo_dir ~obj_dir x in
+  let vo_dir = build_vo_dir ~obj_dir ~prefix:x.prefix in
   let install_vo_dir = String.concat ~sep:"/" x.prefix in
   let native_objs =
     match mode with
     | Coq_mode.Native ->
-      let cmxs_obj = cmxs_of_mod ~wrapper_name x in
-      List.map
-        ~f:(fun x ->
-          ( Path.Build.relative vo_dir x
-          , Filename.(concat (concat install_vo_dir ".coq-native") x) ))
-        cmxs_obj
+      native_obj_files_core x ~wrapper_name ~vo_dir ~install_vo_dir
+    (* split mode doesn't produce any file for coqc *)
+    | Split _
     | VoOnly
     | Legacy ->
       []
@@ -81,6 +95,16 @@ let obj_files x ~wrapper_name ~mode ~obj_dir ~obj_files_mode =
   List.map obj_files ~f:(fun fname ->
       (Path.Build.relative vo_dir fname, Filename.concat install_vo_dir fname))
   @ native_objs
+
+let vo_obj_file x ~obj_dir =
+  let vo_dir = build_vo_dir ~obj_dir ~prefix:x.prefix in
+  let fname = x.name ^ ".vo" in
+  Path.Build.relative vo_dir fname
+
+let native_obj_files x ~wrapper_name ~obj_dir =
+  let vo_dir = build_vo_dir ~obj_dir ~prefix:x.prefix in
+  let install_vo_dir = String.concat ~sep:"/" x.prefix in
+  native_obj_files_core x ~wrapper_name ~vo_dir ~install_vo_dir
 
 let to_dyn { source; prefix; name } =
   let open Dyn.Encoder in