+++ test/blackbox-tests/test-cases/coq/github3624.t/run.t
@@ -12,5 +12,5 @@ that the error message is good when the coq extension is not enabled.
   1 | (coq.theory
   2 |  (name foo))
   Error: 'coq.theory' is available only when coq is enabled in the dune-project
-  file. You must enable it using (using coq 0.3) in your dune-project file.
+  file. You must enable it using (using coq 0.4) in your dune-project file.
   [1]