+++ test/blackbox-tests/test-cases/coq/native-split.t/baz/a/a.v
@@ -1,0 +1,1 @@
+Definition aa := 4.